Make enums.py
